# src/generated/src/main/java/com/example/springtest/api/AuthorApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.2.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.example.springtest.api;

import com.example.springtest.api.dto.request.AuthorCreateRequest;
import com.example.springtest.api.dto.response.AuthorResponse;
import java.util.UUID;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-04-22T16:30:49.788054500+05:00[Asia/Yekaterinburg]")
@Validated
@Tag(name = "Author", description = "the Author API")
public interface AuthorApi {

    /**
     * POST /api/authors : Create author
     *
     * @param authorCreateRequest  (required)
     * @return Author created (status code 200)
     *         or Validation error (status code 400)
     */
    @Operation(
        operationId = "createAuthor",
        summary = "Create author",
        tags = { "Author" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Author created", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = AuthorResponse.class))
            }),
            @ApiResponse(responseCode = "400", description = "Validation error")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/api/authors",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<AuthorResponse> createAuthor(
        @Parameter(name = "AuthorCreateRequest", description = "", required = true) @Valid @RequestBody AuthorCreateRequest authorCreateRequest
    );


    /**
     * DELETE /api/authors/{id} : Delete author
     *
     * @param id  (required)
     * @return Author deleted (status code 204)
     */
    @Operation(
        operationId = "deleteAuthor",
        summary = "Delete author",
        tags = { "Author" },
        responses = {
            @ApiResponse(responseCode = "204", description = "Author deleted")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/api/authors/{id}"
    )
    
    ResponseEntity<Void> deleteAuthor(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") UUID id
    );


    /**
     * GET /api/authors/{id} : Get author by id
     *
     * @param id  (required)
     * @return Author found (status code 200)
     *         or Author not found (status code 404)
     */
    @Operation(
        operationId = "getAuthorById",
        summary = "Get author by id",
        tags = { "Author" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Author found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = AuthorResponse.class))
            }),
            @ApiResponse(responseCode = "404", description = "Author not found")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/authors/{id}",
        produces = { "application/json" }
    )
    
    ResponseEntity<AuthorResponse> getAuthorById(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") UUID id
    );


    /**
     * PUT /api/authors/{id} : Update author
     *
     * @param id  (required)
     * @param authorCreateRequest  (required)
     * @return Author updated (status code 200)
     */
    @Operation(
        operationId = "updateAuthor",
        summary = "Update author",
        tags = { "Author" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Author updated", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = AuthorResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/api/authors/{id}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<AuthorResponse> updateAuthor(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") UUID id,
        @Parameter(name = "AuthorCreateRequest", description = "", required = true) @Valid @RequestBody AuthorCreateRequest authorCreateRequest
    );

}