# src/generated/src/main/java/com/example/springtest/api/ExternalApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.2.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.example.springtest.api;

import com.example.springtest.api.dto.request.ExternalStudentRequest;
import com.example.springtest.api.dto.response.ExternalStudentResponse;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-04-22T16:30:49.788054500+05:00[Asia/Yekaterinburg]")
@Validated
@Tag(name = "External", description = "the External API")
public interface ExternalApi {

    /**
     * POST /external/students : Create external student
     *
     * @param externalStudentRequest  (required)
     * @return Created external student (status code 200)
     */
    @Operation(
        operationId = "createExternalStudent",
        summary = "Create external student",
        tags = { "External" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Created external student", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ExternalStudentResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/external/students",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<ExternalStudentResponse> createExternalStudent(
        @Parameter(name = "ExternalStudentRequest", description = "", required = true) @Valid @RequestBody ExternalStudentRequest externalStudentRequest
    );


    /**
     * GET /external/students/{id} : Get external student info
     *
     * @param id  (required)
     * @return External student info (status code 200)
     */
    @Operation(
        operationId = "getExternalStudent",
        summary = "Get external student info",
        tags = { "External" },
        responses = {
            @ApiResponse(responseCode = "200", description = "External student info", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ExternalStudentResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/external/students/{id}",
        produces = { "application/json" }
    )
    
    ResponseEntity<ExternalStudentResponse> getExternalStudent(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") String id
    );

}