# src/generated/src/main/java/com/example/springtest/api/StudentApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.2.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.example.springtest.api;

import com.example.springtest.api.dto.request.StudentCreateRequest;
import com.example.springtest.api.dto.response.StudentResponse;
import java.util.UUID;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-04-22T16:30:49.788054500+05:00[Asia/Yekaterinburg]")
@Validated
@Tag(name = "Student", description = "the Student API")
public interface StudentApi {

    /**
     * POST /api/students : Create student
     *
     * @param studentCreateRequest  (required)
     * @return Student created (status code 200)
     *         or Validation error (status code 400)
     */
    @Operation(
        operationId = "createStudent",
        summary = "Create student",
        tags = { "Student" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Student created", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = StudentResponse.class))
            }),
            @ApiResponse(responseCode = "400", description = "Validation error")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/api/students",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<StudentResponse> createStudent(
        @Parameter(name = "StudentCreateRequest", description = "", required = true) @Valid @RequestBody StudentCreateRequest studentCreateRequest
    );


    /**
     * DELETE /api/students/{id} : Delete student
     *
     * @param id  (required)
     * @return Student deleted (status code 204)
     */
    @Operation(
        operationId = "deleteStudent",
        summary = "Delete student",
        tags = { "Student" },
        responses = {
            @ApiResponse(responseCode = "204", description = "Student deleted")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/api/students/{id}"
    )
    
    ResponseEntity<Void> deleteStudent(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") UUID id
    );


    /**
     * GET /api/students/{id} : Get student by id
     *
     * @param id  (required)
     * @return Student found (status code 200)
     *         or Student not found (status code 404)
     */
    @Operation(
        operationId = "getStudentById",
        summary = "Get student by id",
        tags = { "Student" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Student found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = StudentResponse.class))
            }),
            @ApiResponse(responseCode = "404", description = "Student not found")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/students/{id}",
        produces = { "application/json" }
    )
    
    ResponseEntity<StudentResponse> getStudentById(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") UUID id
    );


    /**
     * PUT /api/students/{id} : Update student
     *
     * @param id  (required)
     * @param studentCreateRequest  (required)
     * @return Student updated (status code 200)
     */
    @Operation(
        operationId = "updateStudent",
        summary = "Update student",
        tags = { "Student" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Student updated", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = StudentResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/api/students/{id}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<StudentResponse> updateStudent(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") UUID id,
        @Parameter(name = "StudentCreateRequest", description = "", required = true) @Valid @RequestBody StudentCreateRequest studentCreateRequest
    );

}