# src/main/resources/openapi.yaml
openapi: 3.0.3

info:
  title: Education Management API
  version: 1.0.0
  description: API для управления пользователями, авторами и студентами

tags:
  - name: User
  - name: Author
  - name: Student
  - name: External

paths:

  # USER

  /api/users:
    post:
      tags: [ User ]
      summary: Create user
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateRequest'
      responses:
        '200':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/users/{id}:
    get:
      tags: [ User ]
      summary: Get user by id
      operationId: getUserById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags: [ User ]
      summary: Update user
      operationId: updateUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateRequest'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Validation error
        '404':
          description: User not found

    delete:
      tags: [ User ]
      summary: Delete user
      operationId: deleteUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: User deleted
        '404':
          description: User not found

  # AUTHOR

  /api/authors:
    post:
      tags: [ Author ]
      summary: Create author
      operationId: createAuthor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthorCreateRequest'
      responses:
        '200':
          description: Author created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthorResponse'
        '400':
          description: Validation error

  /api/authors/{id}:
    get:
      tags: [ Author ]
      summary: Get author by id
      operationId: getAuthorById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Author found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthorResponse'
        '404':
          description: Author not found

    put:
      tags: [ Author ]
      summary: Update author
      operationId: updateAuthor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthorCreateRequest'
      responses:
        '200':
          description: Author updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthorResponse'

    delete:
      tags: [ Author ]
      summary: Delete author
      operationId: deleteAuthor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Author deleted

  # STUDENT

  /api/students:
    post:
      tags: [ Student ]
      summary: Create student
      operationId: createStudent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentCreateRequest'
      responses:
        '200':
          description: Student created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentResponse'
        '400':
          description: Validation error

  /api/students/{id}:
    get:
      tags: [ Student ]
      summary: Get student by id
      operationId: getStudentById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Student found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentResponse'
        '404':
          description: Student not found

    put:
      tags: [ Student ]
      summary: Update student
      operationId: updateStudent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentCreateRequest'
      responses:
        '200':
          description: Student updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentResponse'

    delete:
      tags: [ Student ]
      summary: Delete student
      operationId: deleteStudent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Student deleted

  # EXTERNAL

  /external/students/{id}:
    get:
      tags: [ External ]
      summary: Get external student info
      operationId: getExternalStudent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: External student info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExternalStudentResponse'

  /external/students:
    post:
      tags: [ External ]
      summary: Create external student
      operationId: createExternalStudent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExternalStudentRequest'
      responses:
        '200':
          description: Created external student
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExternalStudentResponse'

components:
  schemas:

    # USER
    UserCreateRequest:
      type: object
      required: [ username, profile ]
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 100
        profile:
          $ref: '#/components/schemas/ProfileRequest'

    ProfileRequest:
      type: object
      required: [ firstName, lastName ]
      properties:
        firstName:
          type: string
        lastName:
          type: string

    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        profile:
          $ref: '#/components/schemas/ProfileResponse'

    ProfileResponse:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string

    # AUTHOR
    AuthorCreateRequest:
      type: object
      required: [ name, books ]
      properties:
        name:
          type: string
        books:
          type: array
          items:
            $ref: '#/components/schemas/BookCreateRequest'

    BookCreateRequest:
      type: object
      required: [ title ]
      properties:
        title:
          type: string

    AuthorResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        books:
          type: array
          items:
            $ref: '#/components/schemas/BookResponse'

    BookResponse:
      type: object
      properties:
        id:
          type: string
        title:
          type: string

    # STUDENT
    StudentCreateRequest:
      type: object
      required: [ name, lessons ]
      properties:
        name:
          type: string
        lessons:
          type: array
          items:
            $ref: '#/components/schemas/LessonCreateRequest'

    LessonCreateRequest:
      type: object
      required: [ title ]
      properties:
        title:
          type: string

    StudentResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        extra:
          type: string
        lessons:
          type: array
          items:
            $ref: '#/components/schemas/LessonResponse'

    LessonResponse:
      type: object
      properties:
        id:
          type: string
        title:
          type: string

    # EXTERNAL
    ExternalStudentRequest:
      type: object
      required:
        - studentId
        - name
      properties:
        studentId:
          type: string
          format: uuid
        name:
          type: string

    ExternalStudentResponse:
      type: object
      required:
        - studentId
        - extraInfo
      properties:
        studentId:
          type: string
          format: uuid
        extraInfo:
          type: string

    # ERROR
    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
        message:
          type: string
        timestamp:
          type: string
          format: date-time